refactor(plugins): diverse
- reenable tests
- hide property
- gradle warnings
